factor out shared types into shared_types
