+++ src/glyph/mod.rs
@@ -12,6 +12,7 @@ use std::sync::Arc;
 use druid::Data;
 
 use crate::error::{Error, GlifError, GlifErrorInternal};
+use crate::shared_types::{Color, Guideline, Identifier, Line};
 
 /// The name of a glyph.
 pub type GlyphName = Arc<str>;
@@ -105,39 +106,6 @@ pub enum GlifVersion {
 pub struct Advance {
     pub height: f32,
     pub width: f32,
-}
-
-/// Identifiers are optional attributes of several objects in the UFO.
-/// These identifiers are required to be unique within certain contexts
-/// as defined on a per object basis throughout this specification.
-/// Identifiers are specified as a string between one and 100 characters long.
-/// All characters must be in the printable ASCII range, 0x20 to 0x7E.
-#[derive(Debug, Clone, PartialEq)]
-pub struct Identifier(pub(crate) String);
-
-/// A guideline associated with a glyph.
-#[derive(Debug, Clone, PartialEq)]
-pub struct Guideline {
-    /// The line itself.
-    pub line: Line,
-    /// An arbitrary name for the guideline.
-    pub name: Option<String>,
-    /// The color of the line.
-    pub color: Option<Color>,
-    /// Unique identifier for the guideline. This attribute is not required
-    /// and should only be added to guidelines as needed.
-    pub identifier: Option<Identifier>,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum Line {
-    /// A vertical line, passing through a given `x` coordinate.
-    Vertical(f32),
-    /// A horizontal line, passing through a given `y` coordinate.
-    Horizontal(f32),
-    /// An angled line passing through `(x, y)` at `degrees` degrees counteer-clockwise
-    /// to the horizontal.
-    Angle { x: f32, y: f32, degrees: f32 },
 }
 
 #[derive(Debug, Clone, PartialEq)]
@@ -241,15 +209,6 @@ impl std::default::Default for AffineTransform {
     fn default() -> Self {
         Self::identity()
     }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-#[cfg_attr(feature = "druid", derive(Data))]
-pub struct Color {
-    pub red: f32,
-    pub green: f32,
-    pub blue: f32,
-    pub alpha: f32,
 }
 
 #[derive(Debug, Clone, PartialEq)]
